Move Steganography.kt into utils package
